feat: migrate app/js to ESM modules
Replace global-based scripts with ES modules: split the codebase into
focused modules (constants, converter, dom, fetch, page, settings, theme),
add a dynamic-import bootstrap content script, and update vendor files
and manifest accordingly.
