--- app/js/index.js
@@ -1,10 +1,0 @@
-/* global importScripts */
-importScripts('./vendor/md5.js')
-importScripts('./vendor/asciidoctor.js')
-importScripts('./vendor/kroki.js')
-importScripts('./module/namespace.js')
-importScripts('./module/settings.js')
-importScripts('./converter.js')
-importScripts('./background.js')
-importScripts('./vendor/asciidoctor-chart-block-macro.js')
-importScripts('./vendor/asciidoctor-emoji-inline-macro.js')
