+++ app/js/inject.js
@@ -1,59 +1,45 @@
-/* global asciidoctor, webExtension, location */
-asciidoctor.browser.inject = ((webExtension, document, location) => {
-  const Constants = asciidoctor.browser.constants()
-  const Dom = asciidoctor.browser.dom(document)
-  const Settings = asciidoctor.browser.settings(webExtension, Constants)
-  const Theme = asciidoctor.browser.theme(webExtension, Settings, Constants)
-  const Renderer = asciidoctor.browser.renderer(
-    webExtension,
-    document,
-    Constants,
-    Settings,
-    Dom,
-    Theme,
-  )
-  const Converter = asciidoctor.browser.converter(
-    webExtension,
-    Constants,
-    Settings,
-  )
+/* global chrome, browser */
+import { convert } from './module/converter.js'
+import { showError, updateHTML } from './module/page.js'
+import { getBrowserInfo } from './module/settings.js'
 
-  let alreadyRun = false
+const webExtension = typeof browser === 'undefined' ? chrome : browser
 
-  const inject = async (source) => {
-    if (!alreadyRun) {
-      const browserInfo = Settings.getBrowserInfo()
-      if (browserInfo.name === 'Firefox') {
-        let response = {}
-        try {
-          response = await Converter.convert(location.href, source)
-        } catch (e) {
-          response.error = e
-        }
-        showResponse(response)
-      } else {
-        webExtension.runtime.sendMessage(
-          { action: 'convert', source },
-          (response) => {
-            showResponse(response)
-          },
-        )
+let alreadyRun = false
+const inject = async (source) => {
+  if (!alreadyRun) {
+    const browserInfo = getBrowserInfo()
+    if (browserInfo.name === 'Firefox') {
+      let response = {}
+      try {
+        response = await convert(location.href, source)
+      } catch (e) {
+        response.error = e
       }
-      alreadyRun = true
+      showResponse(response)
+    } else {
+      webExtension.runtime.sendMessage(
+        { action: 'convert', source },
+        (response) => {
+          showResponse(response)
+        },
+      )
+    }
+    alreadyRun = true
+  }
+}
+
+const showResponse = (response) => {
+  if (response) {
+    if (response.html) {
+      updateHTML(response).then()
+    } else if (response.error) {
+      showError(response.error)
     }
   }
+}
 
-  const showResponse = (response) => {
-    if (response) {
-      if (response.html) {
-        Renderer.updateHTML(response)
-      } else if (response.error) {
-        Renderer.showError(response.error)
-      }
-    }
-  }
+// eslint-disable-next-line no-unused-vars
+window.inject = inject
 
-  window.inject = inject
-
-  document.getElementById('content').innerHTML += 'Rendering...'
-})(webExtension, document, location)
+document.getElementById('content').innerHTML += 'Rendering...'
