+++ app/js/module/page.js
@@ -1,0 +1,336 @@
+/* global chrome, browser */
+import Constants from './constants.js'
+import {
+  appendOnce,
+  createScriptElement,
+  createStyleElement,
+  createStylesheetLinkElement,
+  decodeEntities,
+  removeElement,
+  replaceStyleElement,
+  replaceStylesheetLinkElement,
+} from './dom.js'
+import { getRenderingSettings, getSetting } from './settings.js'
+import { getDefaultThemeNames, getThemeName } from './theme.js'
+
+const webExtension =
+  typeof browser !== 'undefined'
+    ? browser
+    : typeof chrome !== 'undefined'
+      ? chrome
+      : null
+
+export async function updateHTML(backgroundConverterResponse) {
+  try {
+    removeElement('asciidoctor-browser-custom-js')
+    // Save the scripts that are present at the root of the <body> to be able to restore them after the update
+    // QUESTION: Should we remove this code ? Since using livereload and this extension is not recommended!
+    const scripts = document.body.querySelectorAll(':scope > script')
+    detectLiveReloadJs(scripts)
+    const settings = await getRenderingSettings()
+    const customJavaScript = settings.customScript
+    preprocessing(customJavaScript)
+    await updateBodyHTML(backgroundConverterResponse, scripts)
+    postprocessing(customJavaScript)
+    return true
+  } catch (error) {
+    showError(error)
+    return false
+  }
+}
+
+/**
+ * Update the content of the HTML to show the error
+ * @param error An error
+ */
+export function showError(error) {
+  const message = `${error.name} : ${error.message}`
+  const messageText = `<p>${message}</p>`
+  document.body.innerHTML = `<div id="content"><h4>Error</h4>${messageText}</div>`
+  // eslint-disable-next-line no-console
+  console.error(error.stack)
+}
+
+/**
+ * Append MathJax script
+ */
+const initializeMathJax = (eqnumsValue) => {
+  appendOnce(
+    document.head,
+    createScriptElement({
+      id: 'asciidoctor-mathjax-config',
+      dataset: {
+        eqnumsValue,
+      },
+      src: webExtension.runtime.getURL('js/mathjax/config.js'),
+    }),
+  )
+  appendOnce(
+    document.head,
+    createScriptElement({
+      id: 'asciidoctor-mathjax-initialization',
+      src: webExtension.runtime.getURL('vendor/MathJax-3.2.2/startup.js'),
+      async: true,
+    }),
+  )
+  removeElement('asciidoctor-mathjax-typesetting')
+  /*
+  document.head.appendChild(createScriptElement({
+    id: 'asciidoctor-mathjax-typesetting',
+    src: webExtension.runtime.getURL('js/mathjax/typeset.js')
+  })) */
+}
+
+/**
+ * Append styles
+ * @param doc
+ */
+const appendStyles = (stylesheet) => {
+  // Theme
+  return getThemeName(stylesheet)
+    .then(appendThemeStyle)
+    .then(() => {
+      // Highlight
+      const highlightTheme = 'github'
+      appendOnce(
+        document.head,
+        createStylesheetLinkElement({
+          id: `asciidoctor-browser-${highlightTheme}-highlight-style`,
+          href: webExtension.runtime.getURL(
+            `css/highlight/${highlightTheme}.css`,
+          ),
+        }),
+      )
+    })
+}
+
+/**
+ * @param customJavaScript
+ */
+const preprocessing = (customJavaScript) => {
+  if (customJavaScript && customJavaScript.loadDirective === 'before') {
+    document.head.appendChild(
+      createScriptElement({
+        id: 'asciidoctor-browser-custom-js',
+        innerHTML: customJavaScript.content,
+      }),
+    )
+  }
+}
+
+/**
+ * @param customJavaScript
+ */
+const postprocessing = (customJavaScript) => {
+  if (customJavaScript && customJavaScript.loadDirective === 'after') {
+    document.head.appendChild(
+      createScriptElement({
+        id: 'asciidoctor-browser-custom-js',
+        innerHTML: customJavaScript.content,
+      }),
+    )
+  }
+}
+
+const appendThemeStyle = async (themeName) => {
+  const themeNames = getDefaultThemeNames()
+  // Check if the theme is packaged in the extension... if not it's a custom theme
+  if (themeNames.includes(themeName)) {
+    replaceStylesheetLinkElement(document.head, {
+      id: 'asciidoctor-browser-style',
+      href: webExtension.runtime.getURL(`css/themes/${themeName}.css`),
+    })
+  } else {
+    const customThemeContent = await getSetting(
+      Constants.CUSTOM_THEME_PREFIX + themeName,
+    )
+    if (customThemeContent) {
+      replaceStyleElement(document.head, {
+        id: 'asciidoctor-browser-style',
+        innerHTML: customThemeContent,
+      })
+    }
+  }
+}
+
+/**
+ * Update the <div id="content"> element.
+ * @param html The new HTML content
+ */
+const updateContent = (html) => {
+  const contentElement = document.getElementById('content')
+  if (contentElement) {
+    contentElement.innerHTML = html
+  } else {
+    const contentDiv = document.createElement('div')
+    contentDiv.id = 'content'
+    contentDiv.innerHTML = html
+    document.body.innerHTML = '' // clear <body>
+    document.body.appendChild(contentDiv)
+  }
+}
+
+/**
+ * Update the HTML document with the Asciidoctor document
+ * @param converterResponse The response sent by the converter
+ * @param scripts The scripts to restore
+ */
+const updateBodyHTML = async (converterResponse, scripts) => {
+  const attributes = converterResponse.attributes
+  if (attributes.isFontIcons) {
+    appendFontAwesomeStyle()
+  }
+  await appendStyles(attributes.stylesheet)
+  appendChartistStyle()
+
+  const title = converterResponse.title
+  const doctype = converterResponse.doctype
+  const maxWidth = attributes.maxWidth
+
+  document.title = decodeEntities(title)
+  if (maxWidth) {
+    document.body.style.maxWidth = maxWidth
+  }
+  updateContent(converterResponse.html)
+  let tocClassNames = ''
+  if (
+    attributes.hasSections &&
+    (attributes.tocPosition === 'left' || attributes.tocPosition === 'right')
+  ) {
+    tocClassNames = ` toc2 toc-${attributes.tocPosition}`
+    const tocElement = document.getElementById('toc')
+    if (tocElement !== null) {
+      tocElement.className = 'toc2'
+    }
+  }
+  document.body.className = `${doctype}${tocClassNames}`
+
+  forceLoadDynamicObjects()
+  if (attributes.isStemEnabled) {
+    initializeMathJax(attributes.eqnumsValue)
+  } else {
+    removeElement('asciidoctor-mathjax-config')
+    removeElement('asciidoctor-mathjax-initialization')
+  }
+  appendScripts(scripts)
+  if (attributes.isSourceHighlighterEnabled) {
+    syntaxHighlighting()
+  }
+  drawCharts()
+}
+
+/**
+ * Detect LiveReload.js script to avoid multiple refreshes
+ */
+const detectLiveReloadJs = (scripts) => {
+  let liveReloadDetected = false
+  for (const script of scripts) {
+    if (script.src.indexOf(Constants.LIVERELOADJS_FILENAME) !== -1) {
+      // LiveReload.js detected!
+      liveReloadDetected = true
+      break
+    }
+  }
+  const value = {}
+  value[Constants.LIVERELOADJS_DETECTED_KEY] = liveReloadDetected
+  webExtension.storage.local.set(value)
+}
+
+/**
+ * Append saved scripts
+ */
+const appendScripts = (scripts) => {
+  for (const script of scripts) {
+    if (!isMathTexScript(script)) {
+      document.body.appendChild(script)
+    }
+  }
+}
+
+const isMathTexScript = (script) => {
+  return /math\/tex/i.test(script.type)
+}
+
+/**
+ * Syntax highlighting with Highlight.js
+ */
+const syntaxHighlighting = () => {
+  document.body.querySelectorAll('pre.highlight > code').forEach((node) => {
+    const match = /language-(\S+)/.exec(node.className)
+    if (match !== null && hljs.getLanguage(match[1]) !== null) {
+      hljs.highlightBlock(node)
+    } else {
+      node.className += ' hljs'
+    }
+  })
+}
+
+/**
+ * Draw charts with Chartist
+ */
+const drawCharts = () => {
+  document.body.querySelectorAll('div.ct-chart').forEach((node) => {
+    const options = {
+      height: node.dataset.chartHeight,
+      width: node.dataset.chartWidth,
+      colors: node.dataset.chartColors.split(','),
+    }
+    const dataset = Object.assign({}, node.dataset)
+    const series = Object.values(
+      Object.keys(dataset)
+        .filter((key) => key.startsWith('chartSeries-'))
+        .reduce((obj, key) => {
+          obj[key] = dataset[key]
+          return obj
+        }, {}),
+    ).map((value) => value.split(','))
+    const data = {
+      labels: node.dataset.chartLabels.split(','),
+      series,
+    }
+    Chartist[node.dataset.chartType](node, data, options)
+  })
+}
+
+/**
+ *
+ */
+const appendChartistStyle = () => {
+  appendOnce(
+    document.head,
+    createStylesheetLinkElement({
+      id: 'asciidoctor-browser-chartist-style',
+      href: webExtension.runtime.getURL('css/chartist.min.css'),
+    }),
+  )
+  appendOnce(
+    document.head,
+    createStyleElement({
+      id: 'asciidoctor-browser-chartist-default-style',
+      innerHTML:
+        '.ct-chart .ct-series.ct-series-a .ct-line {stroke:#8EB33B} .ct-chart .ct-series.ct-series-b .ct-line {stroke:#72B3CC} .ct-chart .ct-series.ct-series-a .ct-point {stroke:#8EB33B} .ct-chart .ct-series.ct-series-b .ct-point {stroke:#72B3CC}',
+    }),
+  )
+}
+
+/**
+ *
+ */
+const appendFontAwesomeStyle = () => {
+  appendOnce(
+    document.head,
+    createStylesheetLinkElement({
+      id: 'asciidoctor-browser-font-awesome-style',
+      href: webExtension.runtime.getURL('css/font-awesome.min.css'),
+    }),
+  )
+}
+
+/**
+ * Force dynamic objects to load (iframe, script...)
+ */
+const forceLoadDynamicObjects = () => {
+  document.body.querySelectorAll('iframe').forEach((node) => {
+    node.setAttribute('src', node.getAttribute('src'))
+  })
+}
