+++ app/manifest.json
@@ -239,26 +239,16 @@
         "js/vendor/highlight.js/languages/xquery.min.js",
         "js/vendor/highlight.js/languages/yaml.min.js",
         "js/vendor/highlight.js/languages/zephir.min.js",
-        "js/vendor/asciidoctor.js",
-        "js/vendor/md5.js",
-        "js/vendor/kroki.js",
-        "js/module/namespace.js",
-        "js/module/settings.js",
-        "js/module/dom.js",
-        "js/module/theme.js",
-        "js/loader.js",
-        "js/renderer.js",
-        "js/converter.js",
-        "js/asciidocify.js",
-        "js/vendor/chartist.min.js",
-        "js/vendor/asciidoctor-chart-block-macro.js",
-        "js/vendor/asciidoctor-emoji-inline-macro.js"
+        "js/contentScript.js"
       ]
     }
   ],
   "web_accessible_resources": [
     {
       "resources": [
+        "fonts/fontawesome-webfont.ttf",
+        "fonts/fontawesome-webfont.woff2",
+        "fonts/fontawesome-webfont.woff",
         "css/themes/asciidoctor.css",
         "css/themes/colony.css",
         "css/themes/foundation.css",
@@ -284,19 +274,18 @@
         "img/themes/maker/*.png",
         "img/themes/riak/*.png",
         "img/themes/riak/*.jpg",
-        "js/vendor/kroki.js",
-        "js/vendor/asciidoctor-chart-block-macro.js",
-        "js/vendor/asciidoctor-emoji-inline-macro.js",
-        "js/vendor/asciidoctor.js",
+        "js/module/constants.js",
+        "js/module/converter.js",
+        "js/module/dom.js",
+        "js/module/fetch.js",
+        "js/module/page.js",
+        "js/module/settings.js",
+        "js/module/theme.js",
         "js/vendor/md5.js",
-        "js/loader.js",
-        "js/converter.js",
-        "js/renderer.js",
-        "js/asciidocify.js",
-        "js/vendor/chartist.min.js",
+        "js/vendor/asciidoctor.js",
+        "js/main.js",
         "js/mathjax/config.js",
         "js/mathjax/typeset.js",
-        "fonts/fontawesome-webfont.woff2",
         "vendor/MathJax-3.2.2/*"
       ],
       "matches": [
@@ -335,6 +324,7 @@
     "default_title": "Asciidoctor.js Preview"
   },
   "background": {
-    "service_worker": "js/index.js"
+    "service_worker": "js/background.js",
+    "type": "module"
   }
 }
