+++ package.json
@@ -27,12 +27,12 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@asciidoctor/core": "2.2.6",
+    "@asciidoctor/core": "3.0.4",
     "@biomejs/biome": "^2.4.13",
     "@playwright/test": "^1.59.1",
     "addons-linter": "10.3.0",
     "archiver": "5.3.1",
-    "asciidoctor-kroki": "0.17.0",
+    "asciidoctor-kroki": "0.18.1",
     "bulma": "0.9.4",
     "chartist": "0.11.4",
     "csso": "5.0.5",
